Changed some code organisation related things
