+++ src/opmon/controller/MainMenuCtrl.hpp
@@ -12,7 +12,7 @@ File under GNU GPL v3.0
 #include "../start/Core.hpp"
 #include "../view/MainMenu.hpp"
 #include "AGameScreen.hpp"
-#include "../helper/MenuHelper.hpp"
+#include "MenuHelper.hpp"
 
 namespace OpMon {
     namespace Controller {
